+++ src/commonMain/kotlin/app/cash/turbine/FlowTurbine.kt
@@ -15,6 +15,8 @@
  */
 package app.cash.turbine
 
+import kotlin.time.Duration
+import kotlin.time.ExperimentalTime
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineStart.UNDISPATCHED
 import kotlinx.coroutines.Dispatchers.Unconfined
@@ -26,9 +28,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withTimeout
-import kotlin.time.Duration
-import kotlin.time.ExperimentalTime
-import kotlin.time.seconds
 
 private const val debug = false
 
@@ -48,8 +47,8 @@ private const val debug = false
  * @param timeout Duration each suspending function on [FlowTurbine] will wait before timing out.
  */
 @ExperimentalTime // For timeout.
-suspend fun <T> Flow<T>.test(
-  timeout: Duration = 1.seconds,
+public suspend fun <T> Flow<T>.test(
+  timeout: Duration = Duration.seconds(1),
   validate: suspend FlowTurbine<T>.() -> Unit
 ) {
   coroutineScope {
@@ -109,38 +108,38 @@ suspend fun <T> Flow<T>.test(
  * Represents active collection on a source [Flow] which buffers item emissions, completion,
  * and/or errors as events for consuming.
  */
-interface FlowTurbine<T> {
+public interface FlowTurbine<T> {
   /**
    * Duration that [expectItem], [expectComplete], and [expectError] will wait for an event before
    * throwing a timeout exception.
    */
   @ExperimentalTime
-  val timeout: Duration
+  public val timeout: Duration
 
   /**
    * Cancel collecting events from the source [Flow]. Any events which have already been received
    * will still need consumed using the "expect" functions.
    */
-  suspend fun cancel()
+  public suspend fun cancel()
 
   /**
    * Cancel collecting events from the source [Flow] and ignore any events which have already
    * been received. Calling this function will exit the [test] block.
    */
-  suspend fun cancelAndIgnoreRemainingEvents(): Nothing
+  public suspend fun cancelAndIgnoreRemainingEvents(): Nothing
 
   /**
    * Cancel collecting events from the source [Flow]. Any events which have already been received
    * will be returned.
    */
-  suspend fun cancelAndConsumeRemainingEvents(): List<Event<T>>
+  public suspend fun cancelAndConsumeRemainingEvents(): List<Event<T>>
 
   /**
    * Assert that there are no unconsumed events which have been received.
    *
    * @throws AssertionError if unconsumed events are found.
    */
-  fun expectNoEvents()
+  public fun expectNoEvents()
 
   /**
    * Assert that an event was received and return it.
@@ -148,7 +147,7 @@ interface FlowTurbine<T> {
    *
    * @throws TimeoutCancellationException if no event was received in time.
    */
-  suspend fun expectEvent(): Event<T>
+  public suspend fun expectEvent(): Event<T>
 
   /**
    * Assert that the next event received was an item and return it.
@@ -157,7 +156,7 @@ interface FlowTurbine<T> {
    * @throws AssertionError if the next event was completion or an error.
    * @throws TimeoutCancellationException if no event was received in time.
    */
-  suspend fun expectItem(): T
+  public suspend fun expectItem(): T
 
   /**
    * Assert that the next event received was the flow completing.
@@ -166,7 +165,7 @@ interface FlowTurbine<T> {
    * @throws AssertionError if the next event was an item or an error.
    * @throws TimeoutCancellationException if no event was received in time.
    */
-  suspend fun expectComplete()
+  public suspend fun expectComplete()
 
   /**
    * Assert that the next event received was an error terminating the flow.
@@ -175,20 +174,20 @@ interface FlowTurbine<T> {
    * @throws AssertionError if the next event was an item or completion.
    * @throws TimeoutCancellationException if no event was received in time.
    */
-  suspend fun expectError(): Throwable
+  public suspend fun expectError(): Throwable
 }
 
 private val ignoreRemainingEventsException = CancellationException("Ignore remaining events")
 
-sealed class Event<out T> {
-  object Complete : Event<Nothing>() {
-    override fun toString() = "Complete"
+public sealed class Event<out T> {
+  public object Complete : Event<Nothing>() {
+    override fun toString(): String = "Complete"
   }
-  data class Error(val throwable: Throwable) : Event<Nothing>() {
-    override fun toString() = "Error(${throwable::class.simpleName})"
+  public data class Error(val throwable: Throwable) : Event<Nothing>() {
+    override fun toString(): String = "Error(${throwable::class.simpleName})"
   }
-  data class Item<T>(val value: T) : Event<T>() {
-    override fun toString() = "Item($value)"
+  public data class Item<T>(val value: T) : Event<T>() {
+    override fun toString(): String = "Item($value)"
   }
 }
 
@@ -221,13 +220,13 @@ private class ChannelBasedFlowTurbine<T>(
     cancel()
     return mutableListOf<Event<T>>().apply {
       while (true) {
-        this += events.poll() ?: break
+        this += events.tryReceive().getOrNull() ?: break
       }
     }
   }
 
   override fun expectNoEvents() {
-    val event = events.poll()
+    val event = events.tryReceive().getOrNull()
     if (event != null) {
       unexpectedEvent(event, "no events")
     }
@@ -271,7 +270,7 @@ private class ChannelBasedFlowTurbine<T>(
     val unconsumed = mutableListOf<Event<T>>()
     var cause: Throwable? = null
     while (true) {
-      val event = events.poll() ?: break
+      val event = events.tryReceive().getOrNull() ?: break
       unconsumed += event
       if (event is Event.Error) {
         check(cause == null)